requirements.txt
Django==4.0.6
djangorestframework==3.13.1
psycopg2-binary==2.9.3
Pillow==9.2.0
django-filter==22.1
redis==4.3.4
celery==5.2.7
flower==1.2.0
